move: bn256 point compression
move out of serde, simplify to two-spare-bits
